Add dockerfile for maintenance
Because you need an equivalent deluge installation
and python installation in order to run the pickle
script.  But you have to run that with docker down
or at least everything paused.  This seems simpler to
me:

* bring up kenstir/deluge-maintenance container
* stop kenstir/deluge container
* use kenstir/deluge-maintenance bash to rewrite
  the state file using the python script
* start kenstir/deluge again

The maintenance container is the same without the
service files that start deluge.  The more I look
at it the more I like this lsiobase image.